backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sambanova
pydantic